Fix C4334
```
warning C4334: '&lt;&lt;': result of 32-bit shift implicitly converted to 64 bits (was 64-bit shift intended?)
```

+++ compiler/optimizer/VPHandlers.cpp
@@ -7182,7 +7182,7 @@ static TR::Node *distributeShift(OMR::ValuePropagation *vp, TR::Node *node, int3
 
       if (replaceNode)
          {
-         int64_t pow2Val = (1 << shiftAmount);
+         int64_t pow2Val = int64_t(1) << shiftAmount;
          TR::VPConstraint *lhsORrhs = NULL;
          if (replaceNode == rhsGrandChild) //lhs is the redundant expr
             lhsORrhs = rhsGConstraint;
